consistent naming and exports
